feat: consolidate custom PI code and modularize class functions
